requirements.txt
stashapp-tools
ffmpeg-python
torch
